Updated image

+++ graphics_lib/src/cameras/camera.rs
@@ -5,5 +5,5 @@ use std::fmt::Debug;
 pub trait Camera: Debug {
     // A camera makes a number of rays for a pixel location in the frame
     // The rays get averaged at render time
-    fn rays(&self, x: f32, y: f32) -> Vec<Ray>;
+    fn rays(&self, x: f64, y: f64) -> Vec<Ray>;
 }
